pyproject.toml
[project]
name = "property-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "firecrawl-py>=4.13.0",
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "lxml>=6.0.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
